+++ insts/std/MOVE/BASE.h
@@ -13,11 +13,11 @@ class BASE : public Instrument {
 public:
    BASE();
    virtual ~BASE();
-   virtual int init(float *, int);
+   virtual int init(double *, int);
    virtual int configure();
    virtual int run();
 protected:
-   virtual int localInit(float *, int) = 0;
+   virtual int localInit(double *, int) = 0;
    virtual int finishInit(double, double *) = 0;
    virtual int updatePosition(int) = 0;
    virtual void get_tap(int, int, int, int) = 0;
